+++ ReleaseNotes.md
@@ -62,3 +62,4 @@
 - Add MileSetNtfsCompressionAttributeByHandle function.
 - Add MileGetWofFileCompressionAttributeByHandle function.
 - Add MileSetWofFileCompressionAttributeByHandle function.
+- Add MileRemoveWofFileCompressionAttributeByHandle function.